@@ -4,24 +4,34 @@
   "description": "Migrate tuner-x to v3",
   "scripts": {
     "dev": "vite",
-    "build": "npm run vite-build && npm run rollup-build && npm run post-rollup-build",
+    "build": "npm run vite-build && npm run post-rollup-build",
     "vite-build": "vite build --assetsDir assets",
-    "rollup-build": "rollup src/background/index.js -o dist/background.js -f iife",
     "post-rollup-build": "node build/index.js"
   },
   "dependencies": {
-    "@element-plus/icons-vue": "^0.2.6",
-    "element-plus": "latest",
+    "@element-plus/icons-vue": "^2.1.0",
+    "element-plus": "^2.2.13",
     "moment": "^2.29.4",
-    "vue": "^3.0.5"
+    "pinia": "^2.0.33",
+    "vue": "^3.2.36"
   },
   "devDependencies": {
+    "@iconify-json/ep": "^1.1.4",
+    "@rollup/plugin-typescript": "^11.0.0",
+    "@types/chrome": "^0.0.220",
+    "@types/node": "^18.14.0",
     "@vitejs/plugin-vue": "^4.0.0",
-    "@vue/compiler-sfc": "^3.2.45",
-    "sass": "^1.34.0",
-    "sass-loader": "^11.1.1",
-    "vite": "^4.0.2",
-    "vue-router": "^4.0.2"
+    "esbuild": "0.8.33",
+    "esbuild-windows-64": "0.8.33",
+    "rollup": "^3.18.0",
+    "sass": "^1.53.0",
+    "typescript": "^4.7.2",
+    "unocss": "^0.49.7",
+    "unplugin-vue-components": "^0.24.0",
+    "vite": "^4.1.2",
+    "vite-ssg": "^0.22.1",
+    "vue-router": "^4.0.2",
+    "vue-tsc": "^1.1.3"
   },
   "keywords": [],
   "author": "littlearphone"
