@@ -1,48 +1,53 @@
-(function($) {
-  const key = 'douyu-room'
-  chrome.storage.local.get([key], function(data) {
-    console.log('成功收到Douyu配置信息')
-    const config = { ...data[key] }
-    if (!config.hasOwnProperty('enable')) {
-      config.enable = true
-    }
-    if (!config.hasOwnProperty('fullPage')) {
-      config.fullPage = true
-    }
-    if (!config.hasOwnProperty('danmuSize')) {
-      config.danmuSize = 30
-    }
-    if (!config.enable) {
-      return
-    }
-    (function() {
+(function ($) {
+  if ($('body[douyu]').length) {
+    return console.log('脚本重复注入')
+  }
+  chrome.runtime.onMessage.addListener(function (data, sender, callback) {
+    console.log('Douyu script is on standby')
+    callback({ msg: 'douyu-script-injected' })
+    const config = data.config || {}
+    (function () {
+      if (!config.hasOwnProperty('danMuSize') || config.danMuSize <= 0) {
+        config.danMuSize = 30
+      }
       const innerHTML = `
         .layout-Player #comment-higher-container [class*="danmuItem-"],
         .layout-Player [class*="comment-"] [class*="danmu-"] [class*="danmuItem-"] {
-          font-size: ${config.danmuSize}px !important
+          font-size: ${config.danMuSize}px !important
         }
       `
       $('<style></style>').attr('rel', 'stylesheet').html(innerHTML).appendTo('head')
-    })()
-    $.detect && $.detect('.layout-Player', player => {
-      $.detect('.layout-Player-video', video => {
-        $.detect('#js-player-video', () => {
-          if (!video.find('title-toggle-button').length) {
-            $('<i></i>').appendTo(video).addClass('title-toggle-button')
-              .on('click', () => player.toggleClass('toggle-layout-fold'))
-          }
-          if (!video.find('toolbar-toggle-button').length) {
-            $('<i></i>').appendTo(video).addClass('toolbar-toggle-button')
-              .on('click', () => player.toggleClass('toggle-layout-fold'))
-          }
+    })();
+    (function () {
+      $.detect && $.detect('.layout-Player', player => {
+        $.detect('.layout-Player-video', video => {
+          $.detect('#js-player-video', () => {
+            if (!video.find('title-toggle-button').length) {
+              $('<i></i>').appendTo(video).addClass('title-toggle-button')
+                .on('click', () => player.toggleClass('toggle-layout-fold'))
+            }
+            if (!video.find('toolbar-toggle-button').length) {
+              $('<i></i>').appendTo(video).addClass('toolbar-toggle-button')
+                .on('click', () => player.toggleClass('toggle-layout-fold'))
+            }
+          })
         })
+        player.siblings().hide().parents().each(element => $(element).siblings().hide())
+        if (!config.hasOwnProperty('fullPage') || config.fullPage) {
+          return player.addClass('toggle-layout-fold')
+        }
+        player.removeClass('toggle-layout-fold')
       })
-      player.siblings().hide().parents().each(element => $(element).siblings().hide())
-      if (config.fullPage) {
-        player.addClass('toggle-layout-fold')
-        return
-      }
-      player.removeClass('toggle-layout-fold')
-    })
+    })()
   })
+
+  function initialize () {
+    const $body = $('body')
+    if (!$body.length) {
+      return requestAnimationFrame(initialize)
+    }
+    $body.attr('douyu', '')
+  }
+
+  requestAnimationFrame(initialize)
 })(window.jQuery)
