@@ -1,11 +1,11 @@
-(function($) {
-  function scavenger() {
+(function ($) {
+  function scavenger () {
     const loginModal = $('div').find('.signFlowModal')
     if (loginModal.length) {
       loginModal.parents('div').remove()
     }
-    setTimeout(scavenger, 1000)
+    requestAnimationFrame(scavenger)
   }
 
-  scavenger()
+  requestAnimationFrame(scavenger)
 })(window.jQuery)
